refactor: eliminate type casting anti-pattern, use i64 consistently for SQLite INTEGER
BREAKING CHANGE: Changed all integer types in code intelligence and memory systems
from i32/u32 to i64 to match SQLite's native INTEGER type.

Changes:
- CodeElement: start_line, end_line, complexity_score now i64 (was u32)
- FileAnalysis: complexity_score, test_count now i64 (was u32)
- MemoryEntry: recall_count, token counts, timing metrics now i64 (was i32)
- RepositoryFile: line_count, function_count now i64 (was i32)
- RepoStats: all count fields now i64 (was u32)

Files modified:
- src/prompt/unified_builder.rs - Changed wrapper types to i64
- src/tools/code_fix.rs - Removed database query casts
- src/memory/features/code_intelligence/types.rs - Core struct changes
- src/memory/features/code_intelligence/storage.rs - Removed all i32 casts
- src/memory/features/code_intelligence/parser.rs - Rust parser to i64
- src/memory/features/code_intelligence/typescript_parser.rs - TS parser to i64
- src/memory/features/code_intelligence/javascript_parser.rs - JS parser to i64
- src/git/store.rs - Repository file types to i64
- src/git/client/operations.rs - Line count calculation to i64
- src/memory/core/types.rs - Token counts and metrics to i64
- src/memory/storage/sqlite/core/analysis_operations.rs - Removed recall_count cast
- src/memory/storage/qdrant/mapping.rs - Removed all i32 casts from payload mapping

Rationale:
SQLite's INTEGER type maps to i64 in Rust. Previously we were:
1. Storing i64 in SQLite
2. Retrieving as i64 from sqlx queries
3. Casting to i32/u32 for struct fields
4. Potentially casting back to i64 elsewhere

This caused unnecessary type conversions and potential overflow issues.
Now we use i64 throughout, matching SQLite's native type.

Note: Aggregate functions (COUNT, SUM) still return i32 from sqlx and require
casting to i64. AVG returns i64 and must be cast to f64 for float fields.

+++ src/git/client/operations.rs
@@ -176,7 +176,7 @@ impl GitOperations {
         let content_hash = format!("{:x}", hasher.finish());
         
         let content_str = String::from_utf8_lossy(&content);
-        let line_count = content_str.lines().count() as i32;
+        let line_count = content_str.lines().count() as i64;  // Changed from as i32 to match insert_repository_file signature
         
         // Detect language based on file extension
         let language = if is_rust_file(file_path) {
